backend/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
asyncio==3.4.3
bcrypt==4.2.1
beanie==1.28.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
cryptography==43.0.3
DateTime==5.5
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jwt==1.3.1
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.6.0
Nuitka==2.5.8
ordered-set==4.1.0
packaging==24.2
pluggy==1.5.0
pyarmor==9.0.7
pyarmor.cli.core==7.6.2
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyJWT==2.10.0
pymongo==4.9.2
Pyng==1.6.1
pyparsing==3.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
toml==0.10.2
typer==0.15.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
zope.interface==7.1.1
zstandard==0.23.0
